@@ -8,9 +8,11 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.provider.Settings;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AlertDialog;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
@@ -36,7 +38,7 @@ public class MPermissionUtils {
         requestPermissions(fragment, requestCode, permission, callback);
     }
 
-    public static void requestPermissionsResult(android.support.v4.app.Fragment fragment, int requestCode
+    public static void requestPermissionsResult(Fragment fragment, int requestCode
             , String[] permission, OnPermissionListener callback) {
         requestPermissions(fragment, requestCode, permission, callback);
     }
@@ -68,8 +70,8 @@ public class MPermissionUtils {
 //                } else if (object instanceof android.app.Fragment) {
 //                    ((android.app.Fragment) object).requestPermissions(deniedPermissions
 //                            .toArray(new String[deniedPermissions.size()]), requestCode);
-//                } else if (object instanceof android.support.v4.app.Fragment) {
-//                    ((android.support.v4.app.Fragment) object).requestPermissions(deniedPermissions
+//                } else if (object instanceof Fragment) {
+//                    ((Fragment) object).requestPermissions(deniedPermissions
 //                            .toArray(new String[deniedPermissions.size()]), requestCode);
 //                } else {
 //                    mRequestCode = -1;
@@ -80,8 +82,8 @@ public class MPermissionUtils {
                 } else if (object instanceof android.app.Fragment) {
                     ActivityCompat.requestPermissions(((android.app.Fragment) object).getActivity(),deniedPermissions
                             .toArray(new String[deniedPermissions.size()]), requestCode);
-                } else if (object instanceof android.support.v4.app.Fragment) {
-                    ActivityCompat.requestPermissions(((android.support.v4.app.Fragment) object).getActivity(),deniedPermissions
+                } else if (object instanceof Fragment) {
+                    ActivityCompat.requestPermissions(((Fragment) object).getActivity(),deniedPermissions
                             .toArray(new String[deniedPermissions.size()]), requestCode);
                 } else {
                     mRequestCode = -1;
@@ -97,8 +99,8 @@ public class MPermissionUtils {
         Context context;
         if (object instanceof android.app.Fragment) {
             context = ((android.app.Fragment) object).getActivity();
-        } else if (object instanceof android.support.v4.app.Fragment) {
-            context = ((android.support.v4.app.Fragment) object).getActivity();
+        } else if (object instanceof Fragment) {
+            context = ((Fragment) object).getActivity();
         } else {
             context = (Activity) object;
         }
@@ -188,7 +190,7 @@ public class MPermissionUtils {
         }
 
         boolean isActivity = object instanceof android.app.Activity;
-        boolean isSupportFragment = object instanceof android.support.v4.app.Fragment;
+        boolean isSupportFragment = object instanceof Fragment;
         boolean isAppFragment = object instanceof android.app.Fragment;
 
         if (!(isActivity || isSupportFragment || isAppFragment)) {
