@@ -3,8 +3,9 @@ package com.mark.aoplibrary.utils;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
-import android.support.v4.util.SimpleArrayMap;
+
+import androidx.annotation.NonNull;
+import androidx.collection.SimpleArrayMap;
 
 import com.mark.aoplibrary.MarkAOPHelper;
 
